# shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/request-code": {
      "post": {
        "operationId": "AuthController_requestCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCodeResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Не пройдена проверка капчи"
          },
          "429": {
            "description": "Вход по этому номеру запрошен меньше минуты назад"
          },
          "503": {
            "description": "Провайдер недоступен — войти сейчас нельзя"
          }
        },
        "summary": "Начать вход: мобильная авторизация на указанный номер. Дальше человек либо\nподтверждает вход на SIM-карте, либо получает от провайдера код в SMS —\nчто именно, говорит опрос статуса (`session`), а не этот ответ.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/session": {
      "post": {
        "operationId": "AuthController_pollSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionPollDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusDto"
                }
              }
            }
          }
        },
        "summary": "Узнать, чем кончилась заявка. Фронт опрашивает этот метод, пока идёт\nмобильная авторизация; при подтверждении в ответе сразу токен.\n\nPOST, а не GET: телефон и секрет сессии в query-строке осели бы в логах nginx.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/mobile-id/callback/{secret}": {
      "post": {
        "operationId": "AuthController_mobileIdCallback",
        "parameters": [
          {
            "name": "secret",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Уведомление принято"
          }
        },
        "summary": "Приёмник статусов мобильной авторизации.\n\nТело приходит без подписи, поэтому решения по нему не принимаются: приёмник\nтолько помечает заявку как требующую перепроверки, а статус всё равно\nспрашивается у провайдера в `session`. Иначе тот, кто подберёт id заявки,\nприслал бы «status: 1» и получил чужой вход.\n\nСекрет заявки в адресе — он же `sessionId`, выданный тому, кто вход начал:\nбез него приёмник даже не ищет заявку.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify-code": {
      "post": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверный, истёкший код или исчерпаны попытки"
          }
        },
        "summary": "Обменять код из сообщения на JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Профиль текущего пользователя",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/me/email/request-code": {
      "post": {
        "operationId": "AuthController_requestEmailCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Код отправлен на указанную почту"
          },
          "409": {
            "description": "Почта уже используется другим аккаунтом"
          },
          "429": {
            "description": "Код на эту почту запрошен меньше минуты назад"
          },
          "503": {
            "description": "Не удалось отправить письмо"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Прислать код подтверждения почты уже вошедшему человеку. Нужен потому, что\nотзыв подписывается почтой: на неё уходит решение модератора.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me/email/verify": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверный, истёкший код или исчерпаны попытки"
          },
          "409": {
            "description": "Почта уже используется другим аккаунтом"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Подтвердить почту кодом из письма и привязать её к аккаунту",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me/nickname": {
      "patch": {
        "operationId": "AuthController_setNickname",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetNicknameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "Никнейм уже занят"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Выбрать никнейм; он же подтверждает автоматически выданный при регистрации",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Квартирника",
    "description": "HTTP-контракт бэкенда Квартирника",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestCodeDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "Телефон в российском формате или адрес почты — вход принимает и то и другое,\nчто именно, сервер решает по наличию собачки. Регэксп покрывает обе формы\nсразу: разбить на два поля значило бы заставить фронт угадывать до отправки.",
            "pattern": "/^(\\+7\\d{10}|[^@\\s]+@[^@\\s]+\\.[^@\\s]+)$/",
            "example": "+79991234567"
          },
          "captchaToken": {
            "type": "string",
            "description": "Токен SmartCaptcha. Опционален в схеме, потому что в разработке ключей может\nне быть; когда серверный ключ задан, отсутствие токена даёт 400 в CaptchaService."
          }
        },
        "required": [
          "login"
        ]
      },
      "RequestCodeResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Секрет для опроса статуса входа. Больше в ответе ничего нет намеренно:\nподтвердит человек вход на SIM-карте или провайдер перейдёт на код в SMS —\nна этом шаге ещё неизвестно, это выясняет опрос статуса.",
            "example": "3f2a9c1d4b5e6f708192a3b4c5d6e7f8"
          }
        },
        "required": [
          "sessionId"
        ]
      },
      "SessionPollDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "Телефон в российском формате или адрес почты — вход принимает и то и другое,\nчто именно, сервер решает по наличию собачки. Регэксп покрывает обе формы\nсразу: разбить на два поля значило бы заставить фронт угадывать до отправки.",
            "pattern": "/^(\\+7\\d{10}|[^@\\s]+@[^@\\s]+\\.[^@\\s]+)$/",
            "example": "+79991234567"
          },
          "sessionId": {
            "type": "string",
            "description": "Секрет, выданный при запросе входа: без него статус чужого номера не опросить",
            "pattern": "/^[0-9a-f]{32}$/",
            "example": "3f2a9c1d4b5e6f708192a3b4c5d6e7f8"
          }
        },
        "required": [
          "login",
          "sessionId"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "Телефон, если входили по нему; иначе null",
            "example": "+79991234567"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "description": "Подтверждённая почта; null — ещё не привязана. На неё уходит решение\nмодератора, поэтому без неё отзыв оставить нельзя.",
            "example": "me@example.com"
          },
          "nickname": {
            "type": "string",
            "description": "Публичный ник; им подписаны отзывы",
            "example": "maxim_n"
          },
          "id": {
            "type": "string",
            "description": "Идентификатор пользователя"
          },
          "nicknameConfirmed": {
            "type": "boolean",
            "description": "false — ник сгенерирован автоматически и пользователь его ещё не выбирал.\nФронт по этому флагу показывает обязательный шаг ввода ника."
          },
          "consentAccepted": {
            "type": "boolean",
            "description": "false — пользователь не принимал пользовательское соглашение и не давал\nсогласия на обработку персональных данных. Фронт по этому флагу\nпоказывает блокирующее окно; у аккаунтов, заведённых до появления\nсогласия, флаг тоже false."
          }
        },
        "required": [
          "phone",
          "email",
          "nickname",
          "id",
          "nicknameConfirmed",
          "consentAccepted"
        ]
      },
      "SessionStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "`pending` — ждём подтверждения на SIM-карте; `needs-code` — SIM-PUSH не\nсработал, провайдер прислал код в SMS и надо показать поле ввода;\n`confirmed` — вход состоялся и в ответе есть токен; `failed` — провайдер\nаутентификацию не подтвердил; `expired` — попытка истекла или секрет\nне подошёл. Последние два лечатся новым запросом входа.",
            "enum": [
              "pending",
              "needs-code",
              "confirmed",
              "failed",
              "expired"
            ]
          },
          "accessToken": {
            "type": "string",
            "description": "Только при `confirmed`"
          },
          "user": {
            "description": "Только при `confirmed`",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDto"
              }
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "Телефон в российском формате или адрес почты — вход принимает и то и другое,\nчто именно, сервер решает по наличию собачки. Регэксп покрывает обе формы\nсразу: разбить на два поля значило бы заставить фронт угадывать до отправки.",
            "pattern": "/^(\\+7\\d{10}|[^@\\s]+@[^@\\s]+\\.[^@\\s]+)$/",
            "example": "+79991234567"
          },
          "code": {
            "type": "string",
            "description": "Код из SMS. Длина не фиксирована: код выдаёт провайдер, и в тестовом режиме\nон четырёхзначный (1234), а в боевом может быть длиннее.",
            "pattern": "/^\\d{4,8}$/",
            "example": "1234"
          }
        },
        "required": [
          "login",
          "code"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Bearer-токен для заголовка Authorization"
          },
          "user": {
            "description": "Профиль вошедшего пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "EmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "pattern": "/^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$/",
            "example": "me@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "pattern": "/^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$/",
            "example": "me@example.com"
          },
          "code": {
            "type": "string",
            "description": "Код из письма; свой, шестизначный — в отличие от кода провайдера",
            "pattern": "/^\\d{6}$/",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "SetNicknameDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "Никнейм: буквы, цифры и подчёркивание, 3–20 символов. Пробелов нет\nнамеренно — ник подписывает отзывы и читается как идентификатор.\nЗанятость проверяется без учёта регистра.",
            "pattern": "/^[\\p{L}\\p{N}_]{3,20}$/u",
            "example": "maxim_n",
            "minLength": 3,
            "maxLength": 20
          }
        },
        "required": [
          "nickname"
        ]
      },
...
    }
  }
}